Upload gradle build scan
Motivation:

It must be useful to see test results on a website and share the link
with other folks. Since Gradle enterprise is activated in the Armeria
project, we can upload build scans to https://ge.armeria.dev.

An access token is required to upload build scans. As GitHub secrets are
not exposed to the pull requests from public forks, it is not possible to
directly upload the build scans in a CI build workflow.

@ribafish suggested uploading `build-scan-data` to GitHub Actions
artifacts and then publishing the backup scan data in a sperate workflow
run on Armeria repository.

Modifications:

- Added a new step that uploads Gradle build scan data to GitHub Actions
  artifacts to the CI build.
- Redefined the build job name as a deterministic name.
  - A job name is used to fetch the job result from the JOB API.
  - GitHub does not provide a way to get a job ID or a job name in a workflow,
    - https://github.com/orgs/community/discussions/8945
    - https://stackoverflow.com/questions/71593861/how-to-get-github-actions-matrix-job-name-within-job
- Added `gradle-build-scan.yml` workflow which is triggered when a check
  completes.
  - A sleep time is intentionlly added in the workflow because artifacts
    are not as `check_run` is triggered.
  - Scan results are reported as a comment on a pull request.

Result:

You can now see Gradle build scan of a pull request on https://ge.armeria.dev