Add a stale bot
Motivation:

There are many PRs which are open but stale. Possible reasons for the
stale would be:
- An appropriate solution for a problem could not be found; or
- Maintainers or contributors are busy.
- A PR depends on other PRs or a library patch.

In order to separate and manage old PRs more systematically, it would be
a good idea to add `stale` label to the PR.

For PRs marked as stale, maintainers will determine whether to close the
PR or modify it by themselves.

Modifications:

- Add a stale bot that marks inactive PRs for 60 days
- Auto-close is disabled.
- Don't mark issues as stale.

Result:

A stale bot is working in the Armeria repository on a daily basis.